% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ebs_pure}
\alias{ebs_pure}
\title{Monash EBS PURE publications data}
\format{
A data frame with 7 variables:
\describe{
\item{pure_id}{\code{character}. The unique identifier for the publication in PURE.}
\item{year}{\code{integer}. The year of publication.}
\item{authors}{\code{character}. The authors of the publication.}
\item{title}{\code{character}. The title of the publication.}
\item{journal}{\code{character}. The journal where the publication appeared.}
\item{subtype}{\code{character}. The subtype of the publication.}
\item{bib}{\code{character}. A bibliographic citation in the Harvard format.}
\item{doi}{\code{character}. The DOI of the publication where available.}
}
}
\source{
\url{https://research.monash.edu/en/organisations/econometrics-business-statistics/publications/}
}
\usage{
ebs_pure
}
\description{
This dataset contains publications since 2018, downloaded from PURE on 16 January 2025.
Additional data can be updated using the \code{\link{fetch_pure}()} function.
}
\keyword{datasets}
