% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran.R
\name{fetch_cran}
\alias{fetch_cran}
\title{Fetch packages from CRAN}
\usage{
fetch_cran(author_name)
}
\arguments{
\item{author_name}{A character vector containing the authors' names.}
}
\value{
A data frame returning package name, number of downloads,author names and last update date of the package.
}
\description{
This function searches for CRAN packages by author names.
}
\examples{
\dontrun{
fetch_cran("Michael Lydeamore")
}

}
