% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{fetch_doi}
\alias{fetch_doi}
\title{Fetch article information given a DOI}
\usage{
fetch_doi(doi)
}
\arguments{
\item{doi}{A character vector of DOIs.}
}
\value{
A tibble containing the article information.
}
\description{
Retrieves publications for a given list of DOIs using the CrossRef API.
}
\examples{
\dontrun{
fetch_doi("10.1016/j.ijforecast.2023.10.010")
}
}
