% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pure.R
\name{fetch_pure}
\alias{fetch_pure}
\title{Fetch publications from PURE}
\usage{
fetch_pure(years = 2025:2026)
}
\arguments{
\item{years}{A numeric vector of publication years.}
}
\value{
A data frame containing the data fetched from the PURE API covering
the specified publication years.
}
\description{
To download data from PURE, it is necessary to have access to the
API via Simon Angus and the Astro team \url{https://astro.monash.edu/}.
The API key is stored in the environment variable `PURE_API_KEY`.
You can add it to your environment using \code{\link[usethis]{edit_r_environ}()}.
This is end-point restricted to Monash IP addresses. So either use it on campus
or invoke the VPN before using it off campus.
}
\details{
Publications up to 2024 are available in the \code{\link{ebs_pure}} dataset.
}
\seealso{
\code{\link{ebs_pure}}
}
