% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scholar.R
\name{fetch_scholar}
\alias{fetch_scholar}
\title{Fetch publications from Google Scholar}
\usage{
fetch_scholar(scholar_id)
}
\arguments{
\item{scholar_id}{A character vector of Google Scholar IDs.}
}
\value{
A tibble containing all publications for the specified Google Scholar IDs.
}
\description{
Retrieves publications for given Google Scholar IDs,  and returns them as a tibble.
This function retrieves publications for a given Google Scholar ID and formats them into a structured tibble.
}
\examples{
\dontrun{
fetch_scholar("vamErfkAAAAJ")
}

}
