% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran.R
\name{find_cran_packages}
\alias{find_cran_packages}
\title{get_cran_packages}
\usage{
find_cran_packages(first_name, last_name)
}
\arguments{
\item{first_name}{A character vector representing the authors' first names.}

\item{last_name}{A character vector representing the authors' last names.}
}
\value{
A data frame returning package name, number of downloads,author names and last update date of the package.
}
\description{
This function searches for CRAN packages by a given author's first and last name.
}
\examples{
\dontrun{
get_cran_packages("Michael", "Lydeamore")
}

}
