% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcid.R, R/scholar.R
\name{fetch_orcid}
\alias{fetch_orcid}
\alias{fetch_scholar}
\title{Fetch publications from ORCID}
\usage{
fetch_orcid(orcid_ids)

fetch_scholar(scholar_id)
}
\arguments{
\item{orcid_ids}{A character vector of ORCID IDs.}

\item{scholar_id}{A character vector of Google Scholar IDs.}
}
\value{
A tibble containing all publications for the specified ORCID IDs.

A tibble containing all publications for the specified Google Scholar IDs.
}
\description{
Retrieves publications for given ORCID IDs, and returns them as a tibble.

Retrieves publications for given Google Scholar IDs,  and returns them as a tibble.
This function retrieves publications for a given Google Scholar ID and formats them into a structured tibble.
}
\details{
Before this function can be used, you need to authenticate with the ORCID API.
Follow the instructions at <https://github.com/ropensci-archive/rorcid/blob/master/README-not.md#authentication>.
We have found that the second option (a 2-legged OAuth) is necessary. Then store the token obtained
from `orcid_auth()` in your `.Renviron` file by running `usethis::edit_r_environ()`.
It should be of the form `ORCID_TOKEN=<your token>`.
}
\examples{
\dontrun{
fetch_orcid(c("0000-0003-2531-9408", "0000-0001-5738-1471"))
}

\dontrun{
fetch_scholar("vamErfkAAAAJ")
}

}
