% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-monashdeans.R
\docType{data}
\name{monashdeans}
\alias{monashdeans}
\title{Monash Business School Journal Quality List}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 31 rows and 2 columns.
}
\source{
Monash Business School
}
\usage{
data(monashdeans)
}
\value{
A data frame with 31 observations on the following 2 variables:
\describe{
\item{\code{title}: }{Title of the journal}
\item{\code{rank}: }{Since there are no rankings on the Deans list, this simply returns "Deans List"}
}
}
\description{
This is a dataset that contains the list of the Deans Journal Quality list
used by the Monash Business School.
See \url{https://www.intranet.monash/business/research-services/deans-list} for the latest information.
}
\examples{
library(dplyr)
library(stringr)
monashdeans |>
  filter(str_detect(title, "Statist")) |>
  arrange(rank)
}
\keyword{datasets}
