% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiderorchid-package.R
\docType{package}
\name{spiderorchid-package}
\alias{spiderorchid}
\alias{spiderorchid-package}
\title{spiderorchid: Download and wrangle publication data for Monash EBS academic staff}
\description{
A package to scrape and combine publication data from Google Scholar and ORCID.
}
\author{
\strong{Maintainer}: Rob Hyndman \email{Rob.Hyndman@monash.edu}

Authors:
\itemize{
  \item Sherry Tee \email{stee0006@student.monash.edu}
  \item Parnika Khattri \email{pkha0009@student.monash.edu}
}

}
\keyword{internal}
