% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran.R
\name{fetch_cran}
\alias{fetch_cran}
\title{Fetch packages from CRAN}
\usage{
fetch_cran(
  author_names,
  downloads_from = "2000-01-01",
  downloads_to = Sys.Date()
)
}
\arguments{
\item{author_names}{A character vector containing the authors' names in the form used on CRAN.}

\item{downloads_from}{A date or character string in the format "YYYY-MM-DD" specifying the date
from which to start counting downloads. Default is "2000-01-01".}

\item{downloads_to}{A date or character string in the format "YYYY-MM-DD" specifying the last date
for counting downloads. Default is current date.}
}
\value{
A data frame returning meta data about a package including total downloads
between `downloads_from` and `downloads_to`.
}
\description{
This function searches for CRAN packages by author names. Note that some authors
may use different name variations on CRAN (e.g., "Di Cook" and "Dianne Cook"),
so it may be necessary to call the function
with several variations.
}
\examples{
\dontrun{
cran2024 <- fetch_cran(
   author_names = c("Michael Lydeamore", "Di Cook", "Dianne Cook", "Hyndman"),
   downloads_from = "2024-01-01",
   downloads_to = "2024-12-31"
)
}

}
