% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{journal_ranking}
\alias{journal_ranking}
\title{Find rankings of journals from the Monash Business School, ABDC, CORE, SCImago or ERA2010.}
\usage{
journal_ranking(
  title,
  source = c("monash", "abdc", "era2010", "core", "scimago"),
  fuzzy = TRUE,
  only_best = length(title) > 1,
  ...
)
}
\arguments{
\item{title}{A character vector containing (partial) journal names.}

\item{source}{A character string indicating which ranking data base to use. Default \code{"monash"}.}

\item{fuzzy}{Should fuzzy matching be used. If \code{FALSE}, partial exact matching is used.
Otherwise, full fuzzy matching is used.}

\item{only_best}{If \code{TRUE}, only the best matching journal is returned.}

\item{...}{Other arguments are passed to \code{agrepl} (if \code{fuzzy} is \code{TRUE}), or \code{grepl} otherwise.}
}
\value{
A data frame containing the journal title, rank and source for each matching journal.
}
\description{
Given a list of journal titles, this function will return their ranking from various lists. Data sets used are:
\describe{
\item{\cite{\link{monash}}: }{Monash Business School}
\item{\cite{\link{abdc}}: }{Australian Business Deans' Council}
\item{\cite{\link{era2010}}: }{ERA 2010}
\item{\cite{\link{core}}: }{CORE}
\item{\cite{\link{scimago}}: }{SCImago}
}
This function is used in the \href{https://ebsmonash.shinyapps.io/Journal_Rankings/}{Journal rankings shiny app}.
}
\examples{
# Return ranking for individual journals or conferences
journal_ranking("Annals of Statistics")
journal_ranking("Annals of Statistics", "abdc")
journal_ranking("International Conference on Machine Learning")
journal_ranking("International Conference on Machine Learning", "core")
journal_ranking("R Journal", "scimago", only_best = TRUE)
}
\author{
Rob J Hyndman
}
