% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiderorchid-package.R
\docType{package}
\name{spiderorchid-package}
\alias{spiderorchid}
\alias{spiderorchid-package}
\title{spiderorchid: Download and wrangle publication data for Monash EBS academic staff}
\description{
Provide journal rankings from Monash Business School, ABDC, CORE, Scimago and ERA 2010. Fetch publication details from ORCID, Google Scholar, PURE and from DOIs. Find CRAN packages and download statistics for specified authors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/numbats/spiderorchid/}
  \item \url{https://numbats.github.io/spiderorchid/}
}

}
\author{
\strong{Maintainer}: Rob Hyndman \email{Rob.Hyndman@monash.edu}

Authors:
\itemize{
  \item Michael Lydeamore \email{Michael.Lydeamore@monash.edu}
  \item Sherry Tee \email{stee0006@student.monash.edu}
  \item Parnika Khattri \email{pkha0009@student.monash.edu}
}

}
\keyword{internal}
